Start on the mock saliency map
Command to export depth images from a bag file, so the user can annotate and
then re-build the bag file with a new topic added for the edited images.

Incomplete, basic code is there but the topic insert is not working fully, it
adds the image messages in the right place but they contain no data but they do
have size and encoding, which suggests the file have been read.

Also need to check the conversion of the depth image is correct.